TL updated with new text. Transformation updates.
Changes in templets and schema for TL according to TL word document
delivered.
TL_EN.xml updated with the changes. (Elments and attributes left to do).

xi:include in my version of oxygen needs /2003/ in the path to be /2001/
dont know why but the transformation only works with that change.
Continoued with the update of the transformation to pdf. Still problems
with it writing the tag egXML.
Have come to the conclusion that no multilingual version of the
transformation is needed since we will keep the languages seperate.
Need to make templetets for how to structure translations and then use
the split file Florence created to make the translation look like the EN
files.