Updates following Florence recommendations
In folder tei and tei/eng the version with updates according to updates
delivered are made. This is the master version of TL in english. The
file TL_EN.xml and the olf folder tei/en have been removed for use of
the one following Florence changes and the master files.
In folder split_TL the splitted files and transformation from Florence
is located.
In folder Word the word documents with the changes are located
In folder test_pdf the file used for testing the transformation is
located. This will be updated when all changes are made to the master.