Updated pdf transformation
The transformation is working until the point of examples. So all of the
TL gets to be an pdf but the examples are a mess. he examples only
present the node value. Need to work more with getting it to show the
elementnames. All code and some comments for me in swedish are in the
file. I know the error but need to sit some more with it.