FO Examples transformation
Examples transformation updated so it works.
Updates...txt updated with info.
Taglibrary2pdf_updated.xsl updated regarding the transformation of
examples so it now works.
TL.xml is the whole TL in english. Changes in this file made reagrding
one example written in TEI. For it to work it cant be true TEI,  I
changed the path to it by omitting -c and call the namespace prefix
text.